+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.onboarding.OnboardingActivity"
+            android:name=".ui.onboarding.LogInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -39,11 +39,6 @@
         <meta-data
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
-
-        <activity
-            android:name=".ui.onboarding.LogInActivity"
-            android:exported="true"
-            android:theme="@style/Theme.App.Starting"></activity>
         <activity
             android:name=".ui.onboarding.SignUpActivity"
             android:exported="false" />
